fix loss

+++ models/CapNet.py
@@ -65,7 +65,7 @@ class CapNet(nn.Module):
 
             # make pseudo_labels and masks
             masks = torch.where((soft_labels <= t_b) | (soft_labels >= t_a), 1, 0).float()
-            pseudo_labels = torch.where(soft_labels >= t_a, 1, -1).long()                   
+            pseudo_labels = torch.where(soft_labels >= t_a, 1, -1).float()                
             results = {
                 'logits': lb_logits,
                 'pseudo_lb': pseudo_labels,
